% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dagelijkse_gevallen_mean.R
\name{bereken_gemiddelde_per_land}
\alias{bereken_gemiddelde_per_land}
\title{Bereken het gemiddelde aantal dagelijkse gevallen per land}
\usage{
bereken_gemiddelde_per_land(data)
}
\arguments{
\item{data}{Een data frame met de COVID-19 dataset}
}
\value{
Een data frame met het gemiddelde aantal dagelijkse gevallen per land en wordt gesoreteerd per land van de hoogste naar de laagste sterfgevallen.
}
\description{
Bereken het gemiddelde aantal dagelijkse gevallen per land
}
