% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sterfte_gevallen_berekenen.R
\name{calculate_mortality_rate}
\alias{calculate_mortality_rate}
\title{Sterftegevallen berekenen}
\usage{
calculate_mortality_rate(data)
}
\arguments{
\item{data}{De COVID-19 dataset.}
}
\description{
deze functie berkent de sterftegevallen voor COVID-19 per land.
}
