% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Load_data.R
\name{load_covid_data}
\alias{load_covid_data}
\title{Load COVID-19 Data}
\usage{
load_covid_data(data_path)
}
\arguments{
\item{data_path}{The path to the data file.}
}
\value{
A dataframe containing the COVID-19 data.
}
\description{
This function loads the COVID-19 data from a specified path.
}
