% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/COVID19_plot.R
\name{plot_dagelijkse_gevallen}
\alias{plot_dagelijkse_gevallen}
\title{Maak een plot van de dagelijkse gevallen voor een specifiek land}
\usage{
plot_dagelijkse_gevallen(data, land)
}
\arguments{
\item{data}{Een data frame met de COVID-19 dataset}

\item{land}{Naam van het land waarvoor de plot moet worden gemaakt}
}
\description{
Maak een plot van de dagelijkse gevallen voor een specifiek land
}
