% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vergelijken_COVID19.R
\name{vergelijken_invloed}
\alias{vergelijken_invloed}
\title{Vergelijken COVID-19 invloed tussen twee landen}
\usage{
vergelijken_invloed(data, land1, land2)
}
\arguments{
\item{data}{De COVID-19 dataset.}

\item{land1}{De eerste land voor vergelijken.}

\item{land2}{De tweede land voor vergelijken.}
}
\description{
Deze functie vergelijkt de invloede van COVID-19 tussen twee speciefieke landen.
}
